config image deletion from public dir

+++ model/tools.go
@@ -1,26 +1,53 @@
 package model
 
-import "sync"
+import (
+	"blog/util"
+	"os"
+	"path"
+	"sync"
+)
 
 func createImages(blogId string, images ...string) {
-	wg := sync.WaitGroup{}
+	if images != nil {
+		wg := sync.WaitGroup{}
 
-	const sql = `DELETE FROM "blogImages" WHERE "blogId"=?`
-	if _, err := DB.Exec(sql, blogId); err != nil {
-		panic(err)
+		const sql = `DELETE FROM "blogImages" WHERE "blogId"=?`
+		if _, err := DB.Exec(sql, blogId); err != nil {
+			panic(err)
+		}
+
+		for _, image := range images {
+			wg.Add(1)
+
+			go func(image string) {
+				defer wg.Done()
+				const sql = `INSERT INTO "blogImages" (image, "blogId") VALUES (?, ?)`
+
+				if _, err := DB.Exec(sql, image, blogId); err != nil {
+					panic(err)
+				}
+			}(image)
+		}
+		wg.Wait()
 	}
+}
 
-	for _, image := range images {
-		wg.Add(1)
+func deleteImages(images ...string) {
+	if images != nil {
+		wg := sync.WaitGroup{}
 
-		go func(image string) {
-			defer wg.Done()
-			const sql = `INSERT INTO "blogImages" (image, "blogId") VALUES (?, ?)`
+		for _, image := range images {
+			wg.Add(1)
 
-			if _, err := DB.Exec(sql, image, blogId); err != nil {
-				panic(err)
-			}
-		}(image)
+			go func(image string) {
+				defer wg.Done()
+				imagePath := path.Join(util.PublicDir(), image)
+
+				if err := os.Remove(imagePath); err != nil {
+					panic(err)
+				}
+			}(image)
+		}
+		wg.Wait()
 	}
-	wg.Wait()
 }
